added fixes: - fixed p0 crap - CircleTransform class created
