+++ app/src/main/java/by/itacademy/pvt/menu/MenuActivity.kt
@@ -18,35 +18,40 @@ class MenuActivity : Activity(), View.OnClickListener {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_menu)
 
-        findViewById<View>(R.id.dzButtonId)
+        val dz0Button = findViewById<View>(R.id.dzButtonId)
+        val dz1Button = findViewById<View>(R.id.dz1ButtonId)
+        val dz2Button = findViewById<View>(R.id.dz2ButtonId)
+        val dz2LoginButton = findViewById<View>(R.id.dz2LoginButtonId)
+        val dz3Button = findViewById<View>(R.id.dz3ButtonId)
+
+        dz0Button
             .setOnClickListener {
-                onClick(findViewById(R.id.dzButtonId))
+                onClick(dz0Button)
             }
 
-        findViewById<View>(R.id.dz1ButtonId)
+        dz1Button
             .setOnClickListener {
-                onClick(findViewById(R.id.dz1ButtonId))
+                onClick(dz1Button)
             }
 
-        findViewById<View>(R.id.dz2ButtonId)
+        dz2Button
             .setOnClickListener {
-                onClick(findViewById(R.id.dz2ButtonId))
+                onClick(dz2Button)
             }
 
-        findViewById<View>(R.id.dz2LoginButtonId)
+        dz2LoginButton
             .setOnClickListener {
-                onClick(findViewById(R.id.dz2LoginButtonId))
+                onClick(dz2LoginButton)
             }
 
-        findViewById<View>(R.id.dz3ButtonId)
+        dz3Button
             .setOnClickListener {
-                onClick(findViewById(R.id.dz3ButtonId))
+                onClick(dz3Button)
             }
     }
 
-    override fun onClick(p0: View?) {
-        val id = p0!!.id
-        when (id) {
+    override fun onClick(v: View?) {
+        when (v?.id) {
             R.id.dzButtonId -> {
                 val intent = Intent(this, Dz0Activity::class.java)
                 startActivity(intent)
